Moved Manager and Request Download extensions into a separate Swift file.
